Analysis of impacts of covid on death counts
This is one of my assignments for course STA 442 at the University of Toronto. In this report, I am discussing the impacts of covid-19 on death counts in Quebec, focusing on two age groups: people aged under 50 and people aged over 70. I have uploaded a pdf of the report, a pdf of the R codes, and an RMD file containing the codes.